RAG-based-Chatbot-main/requirements.txt
transformers==4.36.0
torch==2.1.0
accelerate==0.25.0
bitsandbytes==0.41.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
streamlit==1.28.1
gradio==4.0.2
pyttsx3==2.90
speechrecognition==3.10.0
pyaudio==0.2.11
sounddevice==0.4.6
scipy==1.11.3
tiktoken==0.5.1
huggingface-hub==0.19.4 